requirements.txt
numpy==1.23.5
scipy>=1.8,<=1.9.1
scikit-image==0.20.0
matplotlib==3.7.1
imageio==2.31.1
Pillow==10.0.0
pyyaml==6.0
opencv-python==4.7.0.72
einops==0.6.1
mediapipe==0.10.5
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
scikit-learn
pandas
tqdm
tensorboard
